requirements.txt
pytest==24.1.2
selenium==4.20.0
requests==2.31.0
allure-pytest==2.13.5
SQLAlchemy==1.4.50